requirements.txt
numpy==1.21.4opencv-pythonPillow==8.2.0pyparsing==3.0.4python-bidi==0.4.2python-dateutil==2.8.2PyYAML==6.0six==1.16.0imutils==0.5.4matplotlib==3.4.3pytesseract